added messaging features

+++ src/App.js
@@ -10,23 +10,49 @@ import Profile from './pages/Profile';
 import Tournaments from './pages/Tournaments';
 import EditProfile from './pages/EditProfile';
 import Requests from './pages/Requests';
-import Messages from './pages/Messages'; // Import the new Messages page
+import Messages from './pages/Messages';
+import Connections from './pages/Connections'; // Import Connections
 import './App.css';
+import { rtdb } from './firebase'; // Import Realtime Database
+import { ref, onDisconnect, set, serverTimestamp } from 'firebase/database'; // Import RTDB functions
 
 function App() {
   const [darkMode, setDarkMode] = useState(false);
   const [user, setUser] = useState(() => {
     try {
       const savedUser = localStorage.getItem('athlinkoUser');
-      return savedUser ? JSON.parse(savedUser) : null;
+      // Additional check to ensure user object is valid
+      if (savedUser) {
+        const parsedUser = JSON.parse(savedUser);
+        if (parsedUser && parsedUser.sub) {
+          return parsedUser;
+        }
+      }
+      return null;
     } catch (error) {
       return null;
     }
   });
 
   useEffect(() => {
-    if (user) {
+    if (user && user.sub) { // <-- IMPORTANT: Ensure user.sub exists
       localStorage.setItem('athlinkoUser', JSON.stringify(user));
+      
+      const userStatusRef = ref(rtdb, `/status/${user.sub}`);
+      
+      const isOnline = {
+        state: 'online',
+        last_changed: serverTimestamp(),
+      };
+      
+      const isOffline = {
+        state: 'offline',
+        last_changed: serverTimestamp(),
+      };
+
+      set(userStatusRef, isOnline);
+      onDisconnect(userStatusRef).set(isOffline);
+
     } else {
       localStorage.removeItem('athlinkoUser');
     }
@@ -54,6 +80,10 @@ function App() {
             path="/profile/:userId" 
             element={user ? <Profile currentUser={user} setUser={setUser} /> : <Navigate to="/login" replace />} 
           />
+           <Route 
+            path="/profile/:userId/connections"
+            element={user ? <Connections currentUser={user} /> : <Navigate to="/login" replace />}
+          />
           
           <Route path="/tournaments" element={user ? <Tournaments user={user} /> : <Navigate to="/login" replace />} />
           <Route 
@@ -64,7 +94,6 @@ function App() {
             path="/requests"
             element={user ? <Requests currentUser={user} /> : <Navigate to="/login" replace />}
           />
-          {/* Add new routes for messaging */}
           <Route
             path="/messages"
             element={user ? <Messages currentUser={user} /> : <Navigate to="/login" replace />}
